pyproject.toml
[tool.poetry]
name = "adsb-aircraft-tracker"
version = "0.1.0"
description = "ADS-B Aircraft Tracker - Real-time aircraft tracking using 1090 MHz Mode S transponder signals"
authors = ["Josh Frye <joshfrye@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pyrtlsdr = "^0.2.9"
numpy = "^1.21.0"
matplotlib = "^3.5.0"
scipy = "^1.7.0"
setuptools = "^80.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"

# Scripts are not available in package-mode = false
# Use: poetry run python adsb_scanner.py instead

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
